Avoid cast to int
This is purely cosmetic. There cannot be an overflow or truncation there but CodeQL complains.

+++ common/src/main/java/io/xeres/common/id/Id.java
@@ -155,7 +155,7 @@ public final class Id
 					throw new IllegalArgumentException("id has an invalid ascii value: " + number);
 				}
 				number -= 'a';
-				number += 10;
+				number += (byte) 10;
 			}
 			else if (number >= '0')
 			{
